tile: Divide instantiations. This enables parallel compilation.
